move de buttons function in another file
