+++ src/App.css
@@ -4,13 +4,13 @@
   display: flex;
   flex-direction: column;
   width: 50%;
-  max-width: 230px;
+  max-width: 700px;
 }
 
 .screen {
   text-align: right;
-  font-size: 10px;
-  padding: 4% 1%;
+  font-size: 24px;
+  padding: 3% 1%;
   width: 98%;
   color: white;
   background-color: rgb(182, 186, 200);
@@ -23,19 +23,21 @@
 
 .calc-btn {
   padding: 15%;
-  font-size: 6px;
+  font-size: 12px;
   background-color: rgb(249, 250, 250);
-  border: none;
-  border-top: solid 1px whitesmoke;
 }
 
 .operation {
   background-color: orange;
+  border: none;
+  border-top: solid 1px whitesmoke;
 }
 
 .numbers,
 .special {
-  border-left: solid 1px whitesmoke;
+  border: none;
+  border-top: solid 1px lightgray;
+  border-left: solid 1px lightgray;
 }
 
 .btn-0 {
