+++ src/components/Calculator.js
@@ -1,26 +1,4 @@
-const Buttons = () => (
-  <div className="buttons">
-    <button type="button" className="calc-btn btn-ac special">AC</button>
-    <button type="button" className="calc-btn btn-+/- special">+/-</button>
-    <button type="button" className="calc-btn btn-% special">%</button>
-    <button type="button" className="calc-btn btn-÷ operation">÷</button>
-    <button type="button" className="calc-btn btn-7 numbers">7</button>
-    <button type="button" className="calc-btn btn-8 numbers">8</button>
-    <button type="button" className="calc-btn btn-9 numbers">9</button>
-    <button type="button" className="calc-btn btn-× operation">×</button>
-    <button type="button" className="calc-btn btn-4 numbers">4</button>
-    <button type="button" className="calc-btn btn-5 numbers">5</button>
-    <button type="button" className="calc-btn btn-6 numbers">6</button>
-    <button type="button" className="calc-btn btn-- operation">-</button>
-    <button type="button" className="calc-btn btn-1 numbers">1</button>
-    <button type="button" className="calc-btn btn-2 numbers">2</button>
-    <button type="button" className="calc-btn btn-3 numbers">3</button>
-    <button type="button" className="calc-btn btn-+ operation">+</button>
-    <button type="button" className="calc-btn btn-0 numbers">0</button>
-    <button type="button" className="calc-btn btn-. numbers">.</button>
-    <button type="button" className="calc-btn btn-= operation">=</button>
-  </div>
-);
+import Buttons from './CalcButtons';
 
 const Calculator = () => (
   <div className="calculator">
